move AuthenticationController into its own subfolder
